refactor(calendar-date-picker): move preset values and date logic to dedicated module for better organization
