+++ superset-frontend/src/filters/components/Time/CalendarDatePicker.tsx
@@ -15,7 +15,8 @@ import { DateLabel } from 'src/explore/components/controls/DateFilterControl/com
 import { Tooltip } from 'src/components/Tooltip';
 import Icons from 'src/components/Icons';
 import {
-  CALENDAR_DATE_PICKER_PRESET_VALUES,
+  CALENDAR_DATE_PICKER_PRESETS,
+  getPresetDates,
   isCalendarPresetActive,
   normalizeCalendarPresetValue,
 } from './calendarDatePickerPresets';
@@ -85,79 +86,6 @@ const PopoverContainer = styled.div`
     }
   }
 `;
-
-const PRESETS = [
-  { label: t('No filter'), value: NO_TIME_RANGE },
-  { label: t('Today'), value: CALENDAR_DATE_PICKER_PRESET_VALUES.currentDay },
-  {
-    label: t('This week'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.currentWeek,
-  },
-  {
-    label: t('This month'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.currentMonth,
-  },
-  {
-    label: t('This year'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.currentYear,
-  },
-  {
-    label: t('Last week'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.previousWeek,
-  },
-  {
-    label: t('Last month'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.previousMonth,
-  },
-  {
-    label: t('Last year'),
-    value: CALENDAR_DATE_PICKER_PRESET_VALUES.previousYear,
-  },
-];
-
-const getPresetDates = (presetValue: string): [Moment, Moment] | null => {
-  const today = moment().startOf('day');
-  const currentWeekStart = today.clone().startOf('isoWeek');
-  const currentMonthStart = today.clone().startOf('month');
-  const currentYearStart = today.clone().startOf('year');
-
-  switch (presetValue) {
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.currentDay:
-      return [today.clone(), today.clone()];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.currentWeek:
-      return [
-        currentWeekStart.clone(),
-        currentWeekStart.clone().add(1, 'week').subtract(1, 'day'),
-      ];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.currentMonth:
-      return [
-        currentMonthStart.clone(),
-        currentMonthStart.clone().add(1, 'month').subtract(1, 'day'),
-      ];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.currentYear:
-      return [
-        currentYearStart.clone(),
-        currentYearStart.clone().add(1, 'year').subtract(1, 'day'),
-      ];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.previousWeek:
-      return [
-        currentWeekStart.clone().subtract(1, 'week'),
-        currentWeekStart.clone().subtract(1, 'day'),
-      ];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.previousMonth:
-      return [
-        currentMonthStart.clone().subtract(1, 'month'),
-        currentMonthStart.clone().subtract(1, 'day'),
-      ];
-    case CALENDAR_DATE_PICKER_PRESET_VALUES.previousYear:
-      return [
-        currentYearStart.clone().subtract(1, 'year'),
-        currentYearStart.clone().subtract(1, 'day'),
-      ];
-    default:
-      return null;
-  }
-};
 
 export default function CalendarDatePicker({
   value,
@@ -247,7 +175,7 @@ export default function CalendarDatePicker({
       <div className="content">
         <div className="presets">
           <strong>{t('Presets')}</strong>
-          {PRESETS.map(preset => (
+          {CALENDAR_DATE_PICKER_PRESETS.map(preset => (
             <button
               key={preset.value}
               type="button"
